Source Code/requirements.txt
altgraph==0.17.3
asgiref==3.7.2
auto-py-to-exe==2.36.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.15.1
Django==4.2.4
Eel==0.16.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
pefile==2023.2.7
Pillow==10.0.0
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.1.1
pywin32==306
pywin32-ctypes==0.2.2
sqlparse==0.4.4
tk==0.1.0
tzdata==2023.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0
